requirements.txt
google-api-python-client
google-auth
google-cloud-storage
google-auth-oauthlib